Move client/src/graph_utils.js to src/graph_utils/graph_utils.js
