+++ client/src/models/populate_budget_measures.js
@@ -1,4 +1,4 @@
-import { get_client } from '../graphql_utils.js';
+import { get_client } from '../graphql_utils/graphql_utils.js';
 import gql from 'graphql-tag';
 import { log_standard_event } from '../core/analytics.js';
 import { Subject } from './subject.js';
