+++ client/src/models/populate_results.js
@@ -1,4 +1,4 @@
-import { get_client } from '../graphql_utils.js';
+import { get_client } from '../graphql_utils/graphql_utils.js';
 import gql from 'graphql-tag';
 import { log_standard_event } from '../core/analytics.js';
 import {
